Fixes to get dramcore running with SST
1. Segfault due to passing reference of local assingned to *ptr. (What's wrong?)
2. Made req_id a 64-bit integer. Strangely, SST's req_id are garbled and huge.

+++ src/common.h
@@ -73,13 +73,13 @@ class Request {
     public:
         Request(CommandType cmd_type, const Address& addr) :
                 cmd_(Command(cmd_type, addr)), arrival_time_(-1), exit_time_(-1), id_(-1) {}
-        Request(CommandType cmd_type, const Address& addr, long arrival_time, int id) :
+        Request(CommandType cmd_type, const Address& addr, long arrival_time, int64_t id) :
             cmd_(Command(cmd_type, addr)), arrival_time_(arrival_time), exit_time_(-1), id_(id) {}
 
         Command cmd_;
         long arrival_time_;
         long exit_time_;
-        int id_ = 0;
+        int64_t id_ = 0;
 
         int Channel() const { return cmd_.Channel(); }
         int Rank() const { return cmd_.Rank(); }
